pyproject.toml
[tool.poetry]
name = "qa_guru_5_19_api_test"
version = "0.1.0"
description = ""
authors = ["Dmitrii <dalariono1@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
pytest = "^7.3.2"
jsonschema = "^4.17.3"
allure-pytest = "^2.13.2"
curlify = "^2.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
